Created MealPlan CRUD

+++ Entities/Meal.cs
@@ -5,7 +5,7 @@ public class Meal
     public int Id { get; set; }
     public string? Name { get; set; } = null!;
     public TimeSpan? Time { get; set; }
-    public int TotalCalories { get; set; }
+    public double TotalCalories { get; set; }
     
     public int MealPlanId { get; set; }
     public MealPlan MealPlan { get; set; } = null!;
